+++ CHANGELOG.md
@@ -16,7 +16,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 - New function added within flatten_cqc_ratings_job to flatten the new assessment column which is now used by CQC to publish the ratings data.
 - Added current_lsoa21 column to the IND CQC pipeline. This column is now included across all jobs, ensuring it is present the Archive outputs.
-- Added provider name into the merged dataframe within the CQC Coverage job.
+- New function added to merge the old CQC ratings and the new assessment ratings.
 
 
 ### Changed
@@ -49,6 +49,8 @@ All notable changes to this project will be documented in this file.
 
 - Removed recode_unknown_codes_to_null function call at preperation step of assessment data within flatten_cqc_ratings job.
 
+- Migrated Polars validation scripts over to use PointBlank (compatible with >= Python 3.11), so far:
+  - locations_raw
 
 ### Improved