+++ CHANGELOG.md
@@ -6,9 +6,19 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 
 ### Added
-- New function added within flatten_cqc_ratings_job to flatten the new assessment column which is now used by CQC to publish the ratings data.
+- Added provider name into the merged dataframe within the CQC Coverage job.
 
+### Changed
+- Removed all of the old Step Functions now the newer ones are fully operational.
+
+### Improved
+
+## [v2025.08.0] - 09/09/2025
+
+### Added
+- New function added within flatten_cqc_ratings_job to flatten the new assessment column which is now used by CQC to publish the ratings data.
 - Added current_lsoa21 column to the IND CQC pipeline. This column is now included across all jobs, ensuring it is present the Archive outputs.
+- New function added to merge the old CQC ratings and the new assessment ratings.
 
 
 ### Changed
@@ -39,7 +49,10 @@ All notable changes to this project will be documented in this file.
   - upgrading to PySpark 3.5
   - upgrading Glue jobs to 5.0 (default Python version is 3.11)
 
-- Removed all of the old Step Functions now the newer ones are fully operational.
+- Removed recode_unknown_codes_to_null function call at preperation step of assessment data within flatten_cqc_ratings job.
+
+- Migrated Polars validation scripts over to use PointBlank (compatible with >= Python 3.11), so far:
+  - locations_raw
 
 ### Improved